+++ inst/testthat/helper_misc.R
@@ -50,3 +50,18 @@ MeasureDummy = R6Class("MeasureDummy", inherit = MeasureRegr,
     )
   )
 mlr3::mlr_measures$add("dummy", MeasureDummy)
+
+flush_redis = function() {
+  config = redux::redis_config()
+  r = redux::hiredis(config)
+  r$FLUSHDB()
+}
+
+expect_rush_reset = function(rush, type = "kill") {
+  rush$reset(type = type)
+  Sys.sleep(1)
+  keys = rush$connector$command(c("KEYS", "*"))
+  if (!test_list(keys, len = 0)) {
+    stopf("Found keys in redis after reset: %s", keys)
+  }
+}