+++ man-roxygen/param_inst_async.R
@@ -1,0 +1,2 @@
+#' @param inst ([FSelectInstanceAsyncSingleCrit] | [FSelectInstanceAsyncMultiCrit] )\cr
+#'  The feature selection instance.